requirements.txt
ttkbootstrap>=1.0.0
Pillow>=9.0.0
PyMuPDF>=1.19.0
pyperclip>=1.8.2
pygments>=2.10.0
python-dateutil>=2.8.2
openpyxl>=3.0.9
python-magic>=0.4.24
chardet>=4.0.0
Send2Trash>=1.8.0
